% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications_surveymonkey.R
\name{clean_application_colnames}
\alias{clean_application_colnames}
\title{clean_application_colnames}
\usage{
clean_application_colnames(survey_df, lang = "en")
}
\arguments{
\item{survey_df}{tibble. Tibble with the applications.}

\item{lang}{character. Which language was used to collect the applications. Either "en" or "de". Defaults to "en".}
}
\value{
tibble with clean column names
}
\description{
cleans surveymonkey column names. snakecases column names and splits the title into "project_id" and "project_title"
}
