% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{clean_kobo}
\alias{clean_kobo}
\title{convert applications into tibble}
\usage{
clean_kobo(survey_df, project_id)
}
\arguments{
\item{survey_df}{tibble tibble of responses with cleaned column names}

\item{project_id}{character. ID of the project, e.g. 2020-04-ERL.}
}
\value{
data frame containing the responses to the questions
}
\description{
cleans the list of applications into a tibble in long format (one person-application row)
}
