% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_to_mariadb}
\alias{connect_to_mariadb}
\title{connect to MariaDB using environment variables}
\usage{
connect_to_mariadb(
  host = Sys.getenv("DBHOST"),
  dbname = Sys.getenv("DBNAME"),
  user = Sys.getenv("DBUSER"),
  password = Sys.getenv("DBPW")
)
}
\arguments{
\item{host}{character. Defaults to Sys.getenv("DBHOST")}

\item{dbname}{character. Defaults to Sys.getenv("DBNAME")}

\item{user}{character. Defaults to Sys.getenv("DBUSER")}

\item{password}{character. Defaults to Sys.getenv("DBPW")}
}
\description{
connect to MariaDB using environment variables
}
