% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_repo_utils.R
\name{extract_urls}
\alias{extract_urls}
\title{extract urls}
\usage{
extract_urls(txt)
}
\arguments{
\item{txt}{character. text of the github issue.}
}
\value{
character. vector of urls
}
\description{
extract urls
}
