% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{get_discarded_emails}
\alias{get_discarded_emails}
\title{get_discarded_emails}
\usage{
get_discarded_emails(mapping_df, selected_ids)
}
\arguments{
\item{mapping_df}{tibble. Tibble with the mapping from applicant id to name & email}

\item{selected}{numeric. vector with numeric ids of those who were selected from the team.}
}
\value{
email addresses of those who were selected
}
\description{
get_discarded_emails
}
