% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{get_feedback}
\alias{get_feedback}
\title{get_feedback
get feedback for all projects or just a specific one}
\usage{
get_feedback(project_id = NULL, lang = "de")
}
\arguments{
\item{project_id}{character. ID of the project, e.g. ERL-03-2020.}

\item{lang.}{character. Which language was used to collect the feedback. Defaults to "de".}
}
\value{
responses to the feedback survey
}
\description{
get_feedback
get feedback for all projects or just a specific one
}
