% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_repo_utils.R
\name{get_gh_issue_data}
\alias{get_gh_issue_data}
\title{get_gh_issue_data
get data from github issue that matches a given project_id}
\usage{
get_gh_issue_data(project_id)
}
\arguments{
\item{project_id}{character. id of the project, in the form YYYY-mm-ABB where ABB is any three-character, uppercase abbreviation}
}
\value{
list. list with title, number, body and labels sub lists
}
\description{
get_gh_issue_data
get data from github issue that matches a given project_id
}
