% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_kobo}
\alias{get_kobo}
\title{get_kobo}
\usage{
get_kobo(url, token = Sys.getenv("KOBO_CORRELAID_TOKEN"))
}
\arguments{
\item{url}{character. URL to download.}

\item{token}{character. API token. defaults to Sys.getenv("KOBO_CORRELAID_TOKEN")}
}
\description{
load an asset by its url
}
