% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects_repo_utils.R
\name{get_project_status}
\alias{get_project_status}
\title{get_project_status
get status of project from the GitHub projects kanban board}
\usage{
get_project_status(gh_issue_num, github_board_id = "3771798")
}
\arguments{
\item{gh_issue_num}{numeric. number of the github issue.}

\item{github_board_id}{character. internal id of the Project in the GitHub API. Get via the API. Defaults to the ID for the CorrelAid projects board.}
}
\value{
character. status / column of the project in the kanban board
}
\description{
get_project_status
get status of project from the GitHub projects kanban board
}
