% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_surveymonkey}
\alias{get_surveymonkey}
\title{get_surveymonkey}
\usage{
get_surveymonkey(id)
}
\arguments{
\item{id}{character. internal surveymonkey id for the survey}
}
\value{
tibble with answers to the survey
}
\description{
get_surveymonkey
}
