% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_path}
\alias{id_path}
\title{id_path. convert from surveymonkey id to path id}
\usage{
id_path(project_id_surveymonkey)
}
\arguments{
\item{project_id_surveymonkey}{character. project id in surveymonkey form, e.g. CIT-11-2020}
}
\value{
character. project id in path form, e.g. 2020-11-CIT
}
\description{
id_path. convert from surveymonkey id to path id
}
