% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_surveymonkey}
\alias{id_surveymonkey}
\title{id_surveymonkey. convert from path id to surveymonkey id}
\usage{
id_surveymonkey(project_id_path)
}
\arguments{
\item{project_id_path}{character. project id in path form, e.g. 2020-11-CIT}
}
\value{
character. project id in surveymonkey form, e.g. CIT-11-2020
}
\description{
id_surveymonkey. convert from path id to surveymonkey id
}
