% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{load_applications}
\alias{load_applications}
\title{load project applications from Kobo}
\usage{
load_applications(asset_url, project_id = NULL)
}
\arguments{
\item{asset_url}{character. URL of the Kobo asset corresponding to the survey.}

\item{project_id}{character. ID of the project, e.g. ERL-03-2020. Defaults to NULL which means not to filter by project}
}
\value{
data frame containing the responses to the questions
}
\description{
loads the applications for a given project
}
