% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveymonkey.R
\name{load_applications_export}
\alias{load_applications_export}
\title{load project applications from export}
\usage{
load_applications_export(
  export_csv_path,
  project_id_surveymonkey = NULL,
  lang = "en"
)
}
\arguments{
\item{export_csv_path}{character. path to csv export from surveymonkey containing the applications}

\item{project_id_surveymonkey}{character. ID of the project, e.g. ERL-03-2020. Defaults to NULL which means not to filter by project}

\item{lang}{character. Which language was used to collect the applications. Defaults to "en" for the "application for correlaid projects (en)" form. "de" corresponds to "Bewerbungsformular für Projektteams (de)" form.}
}
\value{
data frame containing the responses to the questions
}
\description{
loads the applications for a given project from the csv export
}
