% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{load_projects}
\alias{load_projects}
\title{load all data for all projects in the data subfolder}
\usage{
load_projects()
}
\value{
a list of lists where each element represents a project.
}
\description{
load all data for all projects in the data subfolder
}
