% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{local_chapters}
\alias{local_chapters}
\title{CorrelAid local chapters}
\format{
A data frame with 16 rows and 3 variables:
\describe{
  \item{lc_id}{numeric id}
  \item{lc_name}{name of the local chapter, in snakecase}
  \item{lc_name_full}{human readable of the local chapter}
}
}
\source{
\url{https://github.com/CorrelAid/projects/labels}
}
\usage{
local_chapters
}
\description{
A dataset containing all CorrelAid local chapters, derived from 
the `lc` labels in the CorrelAid projects GitHub repo.
}
\keyword{datasets}
