% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalChapter.R
\name{LocalChapter}
\alias{LocalChapter}
\title{LocalChapter
class for a Local Chapter. internal.
creates a local chapter object}
\description{
LocalChapter
class for a Local Chapter. internal.
creates a local chapter object

LocalChapter
class for a Local Chapter. internal.
creates a local chapter object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocalChapter$new()}}
\item \href{#method-to_tibble}{\code{LocalChapter$to_tibble()}}
\item \href{#method-clone}{\code{LocalChapter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalChapter$new(lc_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lc_name}}{character. the lc_name. see projectutils::local_chapters for available values.
to_tibble}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-to_tibble}{}}}
\subsection{Method \code{to_tibble()}}{
representation of local chapter as a one-row tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalChapter$to_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalChapter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
