% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{creates folder for a new project and adds templates for all the necessary files.}
\usage{
new_project(prefix, year, month, data_folder = "")
}
\arguments{
\item{prefix}{character. 3 character, alphabetic uppercase prefix to be used for the project. Usually the first three letters of the organization name.}

\item{year}{numeric. year the project was started in, e.g. 2019.}

\item{month}{numeric. month the project was started in}

\item{data_folder}{character. path to data folder where the project folder should be created. starts at root of the project as defined by here::here. Defaults to "".}
}
\description{
creates folder for a new project and adds templates for all the necessary files.
}
