% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Organization.R
\name{Organization}
\alias{Organization}
\title{Organization
class representing a CorrelAid Partner Organization}
\description{
Organization
class representing a CorrelAid Partner Organization

Organization
class representing a CorrelAid Partner Organization
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{organization_id}}{character. organization_id name of the organization}

\item{\code{organization_name}}{character. organization_name name of the organization}

\item{\code{website}}{character. website of the organization}

\item{\code{about}}{tibble. Returns or sets the about paragraph of the organization.
create a organization}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Organization$new()}}
\item \href{#method-to_tibble}{\code{Organization$to_tibble()}}
\item \href{#method-get_sql_tables}{\code{Organization$get_sql_tables()}}
\item \href{#method-clone}{\code{Organization$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organization$new(organization_id, organization_name, website = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{organization_id}}{character. ID of the organization. three letter abbreviation, uppercase. e.g. COR for CorrelAid}

\item{\code{organization_name}}{character. name of organization}

\item{\code{website}}{character. url of the website. defaults to NA
to_tibble}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-to_tibble}{}}}
\subsection{Method \code{to_tibble()}}{
returns a one row tibble representation of the Volunteer object.
get_sql_tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organization$to_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sql_tables"></a>}}
\if{latex}{\out{\hypertarget{method-get_sql_tables}{}}}
\subsection{Method \code{get_sql_tables()}}{
a list function that returns tibble for each table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organization$get_sql_tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Organization$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
