% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{Project}
\alias{Project}
\title{Project
class representing the basic properties of a CorrelAid Project}
\description{
Project
class representing the basic properties of a CorrelAid Project

Project
class representing the basic properties of a CorrelAid Project
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{project_id}}{character. id of the project, in the form YYYY-mm-ABB where ABB is any three-character, uppercase abbreviation}

\item{\code{name}}{character. title of the project.}

\item{\code{slug}}{character. Returns or sets the slug (the part of the URL after correlaid.org/projects) for the website. Must not contain spaces and only - as special characters.}

\item{\code{is_public}}{boolean. Returns or sets whether the project is made public on the CorrelAid website. Use with caution!}

\item{\code{start_ym}}{character. Return or set the start of the project, in YYYY-mm form. Does not change the project id.}

\item{\code{end_ym}}{character. Return or set the actual end of the project, in YYYY-mm form.}

\item{\code{end_ym_predicted}}{character. Return or set the predicted end of the project, in YYYY-mm form.}

\item{\code{is_internal}}{boolean. Return or set whether or not the project is an internal project.}

\item{\code{url_git_repo}}{character. Return or set the full URL of the git repository, either GitHub or Gitlab.}

\item{\code{url_gh_issue}}{character. Returns the full URL to the GitHub issue in the projects repository. To set, pass the issue number as an integer.}

\item{\code{url_pad}}{character. Returns or sets the full URL to the CodiMD Pad for the project.}

\item{\code{slack_channel}}{character. Returns or sets the name of the Slack channel for the project.}

\item{\code{status_id}}{integer. Returns or sets the integer corresponding to the status of the project.}

\item{\code{organisation_id}}{integer. Returns or sets the integer corresponding to the partner organization of the project.
create a project}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Project$new()}}
\item \href{#method-clone}{\code{Project$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$new(start_ym, abbreviation, name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_ym}}{character. year and month of the predicted start of the project (~kickoff), in the form YYYY-mm}

\item{\code{abbreviation}}{character. Three character, uppercase abbreviation, usually corresponding to the organization, e.g. COR for CorrelAid.}

\item{\code{name}}{character. Short title of the project.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Project$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
