% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectMember.R
\name{ProjectMember}
\alias{ProjectMember}
\title{ProjectMember
class representing a CorrelAid Project Member}
\description{
ProjectMember
class representing a CorrelAid Project Member

ProjectMember
class representing a CorrelAid Project Member
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{project_id}}{character. id of the project in the form YYYY-mm-{ABB}.}

\item{\code{volunteer_id}}{character. id of the volunteer.}

\item{\code{team_id}}{character. id of the team. Defaults to 1.}

\item{\code{start_active_ym}}{character. start_active_ym when the project member became active}

\item{\code{end_active_ym}}{character. end_active_ym when the project member became active}

\item{\code{behaviour_flag}}{boolean. whether or not there was noteworthy behaviour of the volunteer}

\item{\code{behaviour_description}}{character. description of the behaviour.}

\item{\code{is_public}}{boolean. should the member be included in the project description on the website.}

\item{\code{role_id}}{integer. id of the role of the project member.}

\item{\code{role}}{integer. role of the project member.
create a volunteer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ProjectMember$new()}}
\item \href{#method-to_tibble}{\code{ProjectMember$to_tibble()}}
\item \href{#method-set_role}{\code{ProjectMember$set_role()}}
\item \href{#method-set_behaviour}{\code{ProjectMember$set_behaviour()}}
\item \href{#method-get_sql_tables}{\code{ProjectMember$get_sql_tables()}}
\item \href{#method-clone}{\code{ProjectMember$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMember$new(project_id, volunteer_id, role)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{character. ID of the project.}

\item{\code{volunteer_id}}{integer. ID of the volunteer}

\item{\code{role}}{character. name of the project role the volunteer assumes. Check projectutils::roles for available options.
to_tibble}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-to_tibble}{}}}
\subsection{Method \code{to_tibble()}}{
returns a one row tibble representation of the Volunteer object.
set_role
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMember$to_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_role"></a>}}
\if{latex}{\out{\hypertarget{method-set_role}{}}}
\subsection{Method \code{set_role()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMember$set_role(role)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{character. role of the volunteer. see projectutils::roles for available options.
set_behaviour}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_behaviour"></a>}}
\if{latex}{\out{\hypertarget{method-set_behaviour}{}}}
\subsection{Method \code{set_behaviour()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMember$set_behaviour(flag, description = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flag}}{boolean. value of the flag for offending behaviour.}

\item{\code{description}}{character. description of the offending behaviour. Defaults to NA_character_ , needs to be specified if setting the flag to TRUE.
get_sql_tables}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sql_tables"></a>}}
\if{latex}{\out{\hypertarget{method-get_sql_tables}{}}}
\subsection{Method \code{get_sql_tables()}}{
a list function that returns tibble for each table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMember$get_sql_tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ProjectMember$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
