% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roles}
\alias{roles}
\title{Possible roles for CorrelAid projects}
\format{
A data frame with 3 rows and 2 variables:
\describe{
  \item{role_id}{numeric id}
  \item{role}{name of the role, in snakecase}
}
}
\source{
\url{https://docs.correlaid.org/project-manual/data4good-projects}
}
\usage{
roles
}
\description{
A dataset containing possible roles in CorrelAid projects
}
\keyword{datasets}
