% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{status}
\alias{status}
\title{Possible status for CorrelAid projects}
\format{
A data frame with 10 rows and 3 variables:
\describe{
  \item{status_id}{numeric id}
  \item{status}{human readable name of the status}
  \item{status_description}{description of the status, currently empty}
}
}
\source{
\url{https://github.com/CorrelAid/projects/projects/1}
}
\usage{
status
}
\description{
A dataset containing possible status/phases for CorrelAid projects
taken from the Kanban board in the CorrelAid projects GitHub repo
}
\keyword{datasets}
