% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tag.R
\name{Tag}
\alias{Tag}
\title{Tag
class for a project tag. internal.
creates a tag object}
\description{
Tag
class for a project tag. internal.
creates a tag object

Tag
class for a project tag. internal.
creates a tag object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-new}{\code{Tag$new()}}
\item \href{#method-NA-to_tibble}{\code{Tag$to_tibble()}}
\item \href{#method-unknown-clone}{\code{Tag$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-new"></a>}}
\if{latex}{\out{\hypertarget{method-NA-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$new(category, value = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{character. the tag category. see projectutils::tags for available values.}

\item{\code{value}}{character. the tag value. see projectutils::tags for available values and combinations. Defaults to NA.
to_tibble}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-to_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-NA-to_tibble}{}}}
\subsection{Method \code{to_tibble()}}{
representation of tag as a one-row tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$to_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Tag$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
