% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tags}
\alias{tags}
\title{Possible tags for CorrelAid projects}
\format{
A data frame with 56 rows and 3 variables:
\describe{
  \item{tag_category}{category of the tag}
  \item{tag_value}{value of the tag}
  \item{tag_id}{numeric id}
}
}
\source{
derived from the category:value labels of the GitHub projects repo \url{https://github.com/CorrelAid/projects/labels}
}
\usage{
tags
}
\description{
A dataset containing possible project tags, derived from the GitHub projects repo
}
\keyword{datasets}
