% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R, R/update_project.R
\name{update_project}
\alias{update_project}
\title{updates a top level field of a project}
\usage{
update_project(project, field, value)

update_project(project, field, value)
}
\arguments{
\item{project}{list. the project to be updated.}

\item{field}{character. the field that needs to be updated. only top-level fields work currently}

\item{value}{the new value.}
}
\value{
list. the project.

list. the project.
}
\description{
load all data for all projects in the data subfolder
}
