% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{update_projects_json}
\alias{update_projects_json}
\title{update the list of projects}
\usage{
update_projects_json(data_folder = here::here())
}
\arguments{
\item{data_folder}{character. path to data folder starting at root of the project. defaults to here::here()}
}
\description{
loads the data for all projects and writes the list as json to the data subfolder
}
