% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{use_download_applications}
\alias{use_download_applications}
\title{Use download applications script template}
\usage{
use_download_applications(project_id_path, data_folder = "")
}
\arguments{
\item{project_id_path}{project id in path form, e.g. 2020-11-COR}

\item{data_folder}{character. path to data folder starting at root of the project. defaults to "", i.e. root}
}
\description{
Use download applications script template
}
