% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{use_get_emails}
\alias{use_get_emails}
\title{Use get_application_emails script template}
\usage{
use_get_emails(project_id_path, selected_ids, data_folder = "")
}
\arguments{
\item{project_id_path}{project id in path form, e.g. 2020-11-COR}

\item{selected_ids}{numeric. vector of applicant_id's of selected team members}

\item{data_folder}{character. path to data folder starting at root of the project. defaults to "", i.e. root}
}
\description{
Use get_application_emails script template
}
