% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{use_r6_template}
\alias{use_r6_template}
\title{Use R6 templates}
\usage{
use_r6_template(project_id, data_folder = ".")
}
\arguments{
\item{project_id}{project id in path form, e.g. 2020-11-COR}

\item{data_folder}{character. path to data folder starting at root of the project. defaults to ".", i.e. root}
}
\description{
Use R6 templates
}
