% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{use_team_selection_workflow}
\alias{use_team_selection_workflow}
\title{Use team selection workflow}
\usage{
use_team_selection_workflow(
  project_id_path,
  data_folder = ".",
  export_csv_file = NULL,
  lang = "en"
)
}
\arguments{
\item{project_id_path}{project id in path form, e.g. 2020-11-COR}

\item{data_folder}{character. path to data folder starting at root of the project. defaults to ".", i.e. root}

\item{export_csv_file}{character. file name of csv export file within the team_selection folder. Defaults to NULL, i.e. using the API to get the data.}

\item{lang}{character. Which language was used to collect the applications. Defaults to "en" for the "application for correlaid projects (en)" form. "de" corresponds to "Bewerbungsformular für Projektteams (de)" form.}
}
\description{
Use team selection workflow
}
