% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volunteer.R
\name{Volunteer}
\alias{Volunteer}
\title{Volunteer
class representing a CorrelAid Volunteer}
\description{
Volunteer
class representing a CorrelAid Volunteer

Volunteer
class representing a CorrelAid Volunteer
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{first_name}}{character. first_name of the volunteer}

\item{\code{last_name}}{character. last_name of the volunteer}

\item{\code{email}}{character. Email of the volunteer}

\item{\code{user_gh}}{character. github username of the volunteer.}

\item{\code{user_gl}}{character. gitlab username of the volunteer.}

\item{\code{user_twitter}}{character. twitter username of the volunteer.}

\item{\code{url_website}}{character. personal website of the volunteer.}

\item{\code{url_linkedin}}{character. url to the linkedin profil of the volunteer}

\item{\code{url_xing}}{character. url to the xing profil of the volunteer}

\item{\code{lc_id}}{integer. Returns the integer corresponding to the local chapter of the volunteer.
create a volunteer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Volunteer$new()}}
\item \href{#method-as_tibble}{\code{Volunteer$as_tibble()}}
\item \href{#method-sql_tables}{\code{Volunteer$sql_tables()}}
\item \href{#method-clone}{\code{Volunteer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Volunteer$new(first_name, last_name, email)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{first_name}}{character. first name of the volunteer}

\item{\code{last_name}}{character. last name of the volunteer}

\item{\code{email}}{character. email of the volunteers
as_tibble}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
returns a one row tibble representation of the Volunteer object.
sql_tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Volunteer$as_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sql_tables"></a>}}
\if{latex}{\out{\hypertarget{method-sql_tables}{}}}
\subsection{Method \code{sql_tables()}}{
a list function that returns tibble for each table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Volunteer$sql_tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Volunteer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
