% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{anonymize_applications}
\alias{anonymize_applications}
\title{anonymize_applications}
\usage{
anonymize_applications(survey_df)
}
\arguments{
\item{survey_df}{tibble. Tibble with the applications.}

\item{lang.}{character. Which language was used to collect the applications. Either "en" or "de". Defaults to "en".}
}
\description{
anonymize_applications
}
