% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Description.R
\name{Description}
\alias{Description}
\title{Description
class representing the Description of a CorrelAid project}
\description{
Description
class representing the Description of a CorrelAid project

Description
class representing the Description of a CorrelAid project
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{summary}}{tibble. Returns the summary list of the project. read-only.}

\item{\code{problem}}{tibble. Returns the problem list of the project. read-only.}

\item{\code{data}}{tibble. Returns the data list of the project. read-only.}

\item{\code{approach}}{tibble. Returns the approach list of the project. read-only.}

\item{\code{impact}}{tibble. Returns the impact list of the project. read-only.}

\item{\code{further_links}}{tibble. Returns or sets the further links list of the project.
create a description of a project}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Description$new()}}
\item \href{#method-to_tibble}{\code{Description$to_tibble()}}
\item \href{#method-get_sql_tables}{\code{Description$get_sql_tables()}}
\item \href{#method-clone}{\code{Description$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Description$new(project_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{character. id of the project
to_tibble}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-to_tibble}{}}}
\subsection{Method \code{to_tibble()}}{
returns a one row tibble representation of the Description object.
get_sql_tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Description$to_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sql_tables"></a>}}
\if{latex}{\out{\hypertarget{method-get_sql_tables}{}}}
\subsection{Method \code{get_sql_tables()}}{
function that returns a tibble for each table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Description$get_sql_tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Description$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
