% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Description.R
\name{Description}
\alias{Description}
\title{Description
class representing the Description of a CorrelAid project}
\description{
Description
class representing the Description of a CorrelAid project

Description
class representing the Description of a CorrelAid project
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{summary}}{tibble. Returns the summary list of the project. read-only.}

\item{\code{problem}}{tibble. Returns the problem list of the project. read-only.}

\item{\code{data}}{tibble. Returns the data list of the project. read-only.}

\item{\code{approach}}{tibble. Returns the approach list of the project. read-only.}

\item{\code{impact}}{tibble. Returns the impact list of the project. read-only.}

\item{\code{further_links}}{tibble. Returns or sets the further links list of the project.
create a description of a project}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Description$new()}}
\item \href{#method-add_tag}{\code{Description$add_tag()}}
\item \href{#method-add_local_chapter}{\code{Description$add_local_chapter()}}
\item \href{#method-set_status}{\code{Description$set_status()}}
\item \href{#method-to_tibble}{\code{Description$to_tibble()}}
\item \href{#method-get_sql_tables}{\code{Description$get_sql_tables()}}
\item \href{#method-clone}{\code{Description$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Description$new(project_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_id}}{character. id of the project
add a new tag to the project}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_tag"></a>}}
\if{latex}{\out{\hypertarget{method-add_tag}{}}}
\subsection{Method \code{add_tag()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Description$add_tag(category, value = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{category}}{character. the tag category. see projectutils::tags for available values.}

\item{\code{value}}{character. the tag value. see projectutils::tags for available values and combinations. Defaults to NA.
add a local chapter to the project}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_local_chapter"></a>}}
\if{latex}{\out{\hypertarget{method-add_local_chapter}{}}}
\subsection{Method \code{add_local_chapter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Description$add_local_chapter(lc_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lc_name}}{character. local chapter name. see projectutils::local_chapters for available options.
set status of the project}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_status"></a>}}
\if{latex}{\out{\hypertarget{method-set_status}{}}}
\subsection{Method \code{set_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Description$set_status(status)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{status}}{character. the status. See projectutils::status for options.
to_tibble}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-to_tibble}{}}}
\subsection{Method \code{to_tibble()}}{
returns a one row tibble representation of the Project object.
get_sql_tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Description$to_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_sql_tables"></a>}}
\if{latex}{\out{\hypertarget{method-get_sql_tables}{}}}
\subsection{Method \code{get_sql_tables()}}{
function that returns a tibble for each table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Description$get_sql_tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Description$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
