% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{extract_motivation_questions}
\alias{extract_motivation_questions}
\title{extract_motivation_questions}
\usage{
extract_motivation_questions(survey_df, lang = "en")
}
\arguments{
\item{survey_df}{tibble. Tibble with the applications.}

\item{lang.}{character. Which language was used to collect the applications. Either "en" or "de". Defaults to "en".}
}
\description{
extract_motivation_questions
}
