% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_project.R
\name{load_projects}
\alias{load_projects}
\title{load all data for all projects in the data subfolder}
\usage{
load_projects(data_folder = "")
}
\arguments{
\item{data_folder}{character. path to data folder. starts at root of the project as defined by here::here. Defaults to "".}
}
\value{
a list of lists where each element represents a project.
}
\description{
load all data for all projects in the data subfolder
}
