% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{creates folder for a new project and adds templates for all the necessary files.}
\usage{
new_project(project_id, name, data_folder = here::here())
}
\arguments{
\item{project_id}{character. ID of the project, e.g. 2021-03-COR. see description for details.}

\item{name}{character. short name of the project. defaults to "".}

\item{data_folder}{character. path to data folder starting at root of the project. defaults to "."}
}
\description{
creates folder for a new project and adds templates for all the necessary files.
}
\details{
the project id is composed by the year-month of when the project was started (~when the kickoff took place) and a three-letter, uppercase
abbreviation for the organization. If there are two projects with the same organization starting in the same month, 
two letters of the abbreviation should be used for the organization, and one for the content of the project. e.g. 2021-03-COV for a visualization project
with CorrelAid and 2021-03-COA for an automation project with CorrelAid.
}
