% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R
\name{update_projects_json}
\alias{update_projects_json}
\title{update the list of projects}
\usage{
update_projects_json()
}
\description{
loads the data for all projects and writes the list as json to the data subfolder
and the docs subfolder (the latter to "publish" it).
}
