% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R, R/update_project.R
\name{update_projects_json}
\alias{update_projects_json}
\title{update the list of projects}
\usage{
update_projects_json(data_folder = here::here())

update_projects_json(data_folder = here::here())
}
\arguments{
\item{data_folder}{character. path to data folder where the project folder should be created. starts at root of the project as defined by here::here. Defaults to here::here().}
}
\value{

}
\description{
loads the data for all projects and writes the list as json to the data subfolder
and the docs subfolder (the latter to "publish" it).

loads the data for all projects and writes the list as json to the data subfolder
and the docs subfolder (the latter to "publish" it).
}
