% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{use_fill_project}
\alias{use_fill_project}
\title{Use download applications script template}
\usage{
use_fill_project(project_id_path, data_folder = "")
}
\arguments{
\item{project_id_path}{project id in path form, e.g. 2020-11-COR}

\item{data_folder}{character. path to data folder starting at root of the project. defaults to "", i.e. root}
}
\description{
Use download applications script template
}
