% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{use_team_selection_workflow}
\alias{use_team_selection_workflow}
\title{Use team selection workflow}
\usage{
use_team_selection_workflow(project_id_path, data_folder = ".")
}
\arguments{
\item{project_id_path}{project id in path form, e.g. 2020-11-COR}

\item{data_folder}{character. path to data folder starting at root of the project. defaults to ".", i.e. root}
}
\description{
Use team selection workflow
}
