% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volunteer.R
\name{Volunteer}
\alias{Volunteer}
\title{Volunteer
class representing a CorrelAid Volunteer}
\description{
Volunteer
class representing a CorrelAid Volunteer

Volunteer
class representing a CorrelAid Volunteer
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{first_name}}{character. first_name of the volunteer}

\item{\code{last_name}}{character. last_name of the volunteer}

\item{\code{email}}{character. Email of the volunteer}

\item{\code{user_gh}}{character. github username of the volunteer.}

\item{\code{user_gl}}{character. gitlab username of the volunteer.}

\item{\code{user_twitter}}{character. twitter username of the volunteer.}

\item{\code{url_website}}{character. personal website of the volunteer.}

\item{\code{url_linkedin}}{character. url to the linkedin profil of the volunteer}

\item{\code{url_xing}}{character. url to the xing profil of the volunteer}

\item{\code{lc_id}}{integer. Returns the integer corresponding to the local chapter of the volunteer.
create a volunteer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Volunteer-new}{\code{Volunteer$new()}}
\item \href{#method-Volunteer-to_tibble}{\code{Volunteer$to_tibble()}}
\item \href{#method-Volunteer-set_local_chapter}{\code{Volunteer$set_local_chapter()}}
\item \href{#method-Volunteer-get_sql_tables}{\code{Volunteer$get_sql_tables()}}
\item \href{#method-Volunteer-clone}{\code{Volunteer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Volunteer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Volunteer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Volunteer$new(
  first_name,
  last_name,
  email,
  user_gh = NA_character_,
  user_gl = NA_character_,
  user_twitter = NA_character_,
  url_website = NA_character_,
  url_linkedin = NA_character_,
  url_xing = NA_character_,
  lc_name = NA_character_
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{first_name}}{character. first name of the volunteer}

\item{\code{last_name}}{character. last name of the volunteer}

\item{\code{email}}{character. email of the volunteer}

\item{\code{user_gh}}{character. GitHub username of the volunteer. defaults to NA}

\item{\code{user_gl}}{character. GitLab username of the volunteer. defaults to NA}

\item{\code{user_twitter}}{character. Twitter username of the volunteer. defaults to NA}

\item{\code{url_website}}{character. Personal website of the volunteer. defaults to NA}

\item{\code{url_linkedin}}{character. URL to Linkedin profile of the volunteer. defaults to NA}

\item{\code{url_xing}}{character. URL to Xing profile of the volunteer. defaults to NA}

\item{\code{lc_name}}{character. Name of the local chapter of the volunteer. defaults to NA
to_tibble}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Volunteer-to_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-Volunteer-to_tibble}{}}}
\subsection{Method \code{to_tibble()}}{
returns a one row tibble representation of the Volunteer object.
set_local_chapter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Volunteer$to_tibble()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Volunteer-set_local_chapter"></a>}}
\if{latex}{\out{\hypertarget{method-Volunteer-set_local_chapter}{}}}
\subsection{Method \code{set_local_chapter()}}{
sets the local chapter of the volunteer
get_sql_tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Volunteer$set_local_chapter(lc_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lc_name}}{name of the local chapter. see projectutils::local_chapters for options.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Volunteer-get_sql_tables"></a>}}
\if{latex}{\out{\hypertarget{method-Volunteer-get_sql_tables}{}}}
\subsection{Method \code{get_sql_tables()}}{
a list function that returns tibble for each table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Volunteer$get_sql_tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Volunteer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Volunteer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Volunteer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
