% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_project.R
\name{write_project}
\alias{write_project}
\title{writes back meta.json for a project.}
\usage{
write_project(project, data_folder = here::here())
}
\arguments{
\item{project}{the project.}

\item{data_folder.}{data folder path. defaults to here::here()}
}
\value{
the project (invisibly)
}
\description{
writes back meta.json for a project.
}
