% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_utils.R, R/update_project.R
\name{write_project}
\alias{write_project}
\title{writes back meta.json for a project.}
\usage{
write_project(project, data_folder = "")

write_project(project, data_folder = "")
}
\arguments{
\item{project}{the project.}

\item{data_folder}{character. path to data folder where the project folder is. starts at root of the project as defined by here::here. Defaults to "".}

\item{project.}{the project.}

\item{data_folder.}{data folder path. defaults to here::here()}
}
\value{
the project (invisibly)

the project (invisibly)
}
\description{
writes back meta.json for a project.

writes back meta.json for a project.
}
